# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cricket Knowledge RAG Chatbot API",
    "version": "1.0.0",
    "description": "API documentation for the Cricket Knowledge RAG (Retrieval-Augmented Generation) Chatbot. This backend parses documents (PDF, DOCX, TXT), chunks them, embeds them with the Gemini API, stores them in MongoDB Atlas, performs Vector Search, manages phone authentication, and provides personalized quizzes with score points."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/upload": {
      "post": {
        "summary": "Upload and index a Cricket document",
        "description": "Uploads a PDF, DOCX, or TXT file. The server extracts the raw text, splits it into chunks of 500-1000 characters (with 100 character overlap), generates 768-dimensional embeddings using Gemini, and saves them to MongoDB Atlas for vector search.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "The file to upload (PDF, DOCX, or TXT format)."
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Document parsed, embedded, and indexed successfully."
          },
          "400": {
            "description": "Bad Request."
          },
          "500": {
            "description": "Internal Server Error."
          }
        }
      }
    },
    "/ask/history": {
      "get": {
        "summary": "Get authenticated user's chat history grouped by session",
        "description": "Retrieves all previous Q&A interactions grouped under their respective sessionId, sorted descending by date.",
        "responses": {
          "200": {
            "description": "Successfully retrieved grouped chat sessions.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "sessionId": { "type": "string", "example": "session_123" },
                      "timestamp": { "type": "string", "format": "date-time" },
                      "firstQuestion": { "type": "string", "example": "What is a googly?" },
                      "messages": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "_id": { "type": "string" },
                            "question": { "type": "string" },
                            "answer": { "type": "string" },
                            "timestamp": { "type": "string", "format": "date-time" }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized token."
          }
        }
      }
    },
    "/ask": {
      "post": {
        "summary": "Ask a cricket question (Authenticated)",
        "description": "Queries the vector database using semantic search, constructs context, generates a response from Gemini, and saves the interaction to the user's history under an optional sessionId.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "question": {
                    "type": "string",
                    "example": "Why is a googly difficult to play?"
                  },
                  "sessionId": {
                    "type": "string",
                    "example": "session_123",
                    "description": "Optional unique identifier for grouping questions into conversational threads."
                  },
                  "history": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "role": { "type": "string", "example": "user" },
                        "content": { "type": "string", "example": "Hello" }
                      }
                    }
                  }
                },
                "required": [
                  "question"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Answer generated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "answer": { "type": "string" },
                    "historyId": { "type": "string", "example": "6a50bdf99f755..." },
                    "sources": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "filename": { "type": "string" },
                          "chunkId": { "type": "integer" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}